test all pass , legacy deleted
