--- src/legacy/JuniorCampsPage.jsx
@@ -1,5 +1,0 @@
-export default function JuniorCampsPage() {
-  return (
-    <p>this page is showing camps for junior dev</p>
-  );
-}
