--- src/legacy/NewbieCampsPage.jsx
@@ -1,5 +1,0 @@
-export default function NewbieCampsPage() {
-  return (
-    <p>this page is showing camps for newbie dev</p>
-  );
-}
