--- src/legacy/events/Event.jsx
@@ -1,36 +1,0 @@
-import styled from 'styled-components';
-
-const Image = styled.img`
-    width:15vw;
-    height:15vw;
-  `;
-
-export default function Event({ event, onClickEvent }) {
-  const {
-    id, title, date, imgUrl,
-    hashTags,
-  } = event;
-
-  const handleClickEvent = () => {
-    const tempId = 1;
-    onClickEvent({ tempId });
-  };
-
-  return (
-    <button type="button" onClick={handleClickEvent}>
-      <div><Image src={imgUrl} alt="event-img" /></div>
-      <div>
-        {date}
-      </div>
-      <h3>{title}</h3>
-      <ul>
-        {hashTags.map((hashTag) => (
-          <li key={`${id}-${hashTag}`}>
-            #
-            {hashTag}
-          </li>
-        ))}
-      </ul>
-    </button>
-  );
-}
