--- src/legacy/talks/Talks.jsx
@@ -1,17 +1,0 @@
-import FreeTalk from './Talk';
-
-export default function FreeTalks({
-  freeTalks, onClickFreeTalk,
-}) {
-  return (
-    <ul>
-      {freeTalks.map((freeTalk) => (
-        <FreeTalk
-          key={freeTalk.id}
-          freeTalk={freeTalk}
-          onClickFreeTalk={onClickFreeTalk}
-        />
-      ))}
-    </ul>
-  );
-}
