--- src/legacy/useStudyFormStore.js
@@ -1,15 +1,0 @@
-import { useEffect } from 'react';
-import { studyFormStore } from '../stores/StudyFormStore';
-import useForceUpdate from './useForceUpdate';
-
-export default function useStudyFormStore() {
-  const forceUpdate = useForceUpdate();
-
-  useEffect(() => {
-    studyFormStore.subscribe(forceUpdate);
-
-    return () => studyFormStore.unsubscribe(forceUpdate);
-  }, [forceUpdate]);
-
-  return studyFormStore;
-}
